new configuration includes http endpoint config file
